# Backend/src/main/java/com/intelliedu/intelliedu/dto/EventDto.java
package com.intelliedu.intelliedu.dto;

import java.time.LocalDate;
import java.time.LocalTime;
import java.util.UUID;

import com.fasterxml.jackson.annotation.JsonFormat;
import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonInclude.Include;
import com.intelliedu.intelliedu.config.EventType;

import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

/**
 * EventDto
 */
@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
@JsonInclude(Include.NON_NULL)
public class EventDto {

	private UUID id;

  @NotNull(message = "Name must not be empty")
	private String name;

  @NotNull(message = "Date must not be empty")
  @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "d/M/yyyy")
	private LocalDate date;

  @NotNull(message = "Time must not be empty")
  @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "HH:mm")
	private LocalTime time;
	
	private boolean urgent;

	private boolean important;

	private EventType eventType;

	private String description;

  private boolean shared;
}